--- src/orchestration/prefect_viadot/__init__.py
@@ -1,1 +1,0 @@
-"""prefect-viadot."""
