+++ src/viadot/orchestration/prefect/flows/sharepoint_to_adls.py
@@ -1,6 +1,7 @@
 """Flows for pulling data from/into Sharepoint."""
 
-from orchestration.prefect_viadot.tasks import df_to_adls, sharepoint_to_df
+from viadot.orchestration.prefect.tasks import df_to_adls, sharepoint_to_df
+
 from prefect import flow
