+++ src/viadot/orchestration/prefect/flows/sharepoint_to_databricks.py
@@ -3,7 +3,8 @@
 import contextlib
 from typing import Literal
 
-from orchestration.prefect_viadot.tasks import df_to_databricks, sharepoint_to_df
+from viadot.orchestration.prefect.tasks import df_to_databricks, sharepoint_to_df
+
 from prefect import flow
