+++ src/viadot/orchestration/prefect/flows/transform.py
@@ -3,8 +3,9 @@
 import os
 import shutil
 
-from orchestration.prefect_viadot.tasks import clone_repo, dbt_task
-from orchestration.prefect_viadot.utils import get_credentials
+from viadot.orchestration.prefect.tasks import clone_repo, dbt_task
+from viadot.orchestration.prefect.utils import get_credentials
+
 from prefect import flow, task
