+++ src/viadot/orchestration/prefect/flows/transform_and_catalog.py
@@ -4,8 +4,9 @@ import shutil
 from pathlib import Path
 from typing import Literal
 
-from orchestration.prefect_viadot.tasks import clone_repo, dbt_task, luma_ingest_task
-from orchestration.prefect_viadot.utils import get_credentials
+from viadot.orchestration.prefect.tasks import clone_repo, dbt_task, luma_ingest_task
+from viadot.orchestration.prefect.utils import get_credentials
+
 from prefect import allow_failure, flow, task
