+++ src/viadot/orchestration/prefect/tasks/adls.py
@@ -4,8 +4,9 @@ import contextlib
 from typing import Any
 
 import pandas as pd
-from orchestration.prefect_viadot.exceptions import MissingSourceCredentialsError
-from orchestration.prefect_viadot.utils import get_credentials
+from viadot.orchestration.prefect.exceptions import MissingSourceCredentialsError
+from viadot.orchestration.prefect.utils import get_credentials
+
 from prefect import task
 
 with contextlib.suppress(ImportError):
