+++ src/viadot/orchestration/prefect/tasks/luma.py
@@ -5,7 +5,8 @@ import os
 from pathlib import Path
 from typing import Any, Literal
 
-from orchestration.prefect_viadot.utils import shell_run_command
+from viadot.orchestration.prefect.utils import shell_run_command
+
 from prefect import task
 from prefect.logging import get_run_logger
