+++ src/viadot/orchestration/prefect/tasks/sharepoint.py
@@ -3,11 +3,12 @@
 from typing import Any
 
 import pandas as pd
-from orchestration.prefect_viadot.exceptions import MissingSourceCredentialsError
-from orchestration.prefect_viadot.utils import get_credentials
-from prefect import get_run_logger, task
+from viadot.orchestration.prefect.exceptions import MissingSourceCredentialsError
+from viadot.orchestration.prefect.utils import get_credentials
 from viadot.sources import Sharepoint
 from viadot.sources.sharepoint import SharepointCredentials
+
+from prefect import get_run_logger, task
 
 
 @task(retries=3, retry_delay_seconds=10, timeout_seconds=60 * 60)
